Move FallbackDependenciesRepository to coursier module
